Token service implemented

+++ src/main/java/com/org/cap/controller/AccountRegistrationController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -30,51 +31,89 @@ public class AccountRegistrationController {
 	AccountService accountService;
 
 	@PostMapping(value = "/create")
-	public AccountRegistrationResponse createAccount(@RequestBody AccountRegistration accountInputForm) {
+	public AccountRegistrationResponse createAccount(@RequestBody AccountRegistration accountInputForm, @RequestHeader String authorization){
 		logger.error("In account Registration");
+		
 		AccountRegistrationResponse accountRegistrationResponse = new AccountRegistrationResponse();
-		accountRegistrationResponse.setSuccess(false);
-		if (null != accountInputForm) {
-
+		String isvalidToken = accountService.isTokenValid(authorization);
+		if (isvalidToken.equalsIgnoreCase("success")) {
 			accountRegistrationResponse = accountService.createAccount(accountInputForm);
+			accountRegistrationResponse.setSuccess(true);
+		}else {
+			accountRegistrationResponse.setExceptionDetails("Token cannot be validated");
+			accountRegistrationResponse.setSuccess(false);
 		}
-
 		return accountRegistrationResponse;
 
 	}
 	
 	@GetMapping(value = "/accounts")
-	public List<Account> getAllAccounts() {
+	public AccountRegistrationResponse getAllAccounts(@RequestHeader String authorization){
 		logger.error("In find All Accounts");
+		
+		AccountRegistrationResponse accountRegistrationResponse = new AccountRegistrationResponse();
+		String isvalidToken = accountService.isTokenValid(authorization);
+		if (isvalidToken.equalsIgnoreCase("success")) {
 		List<Account> accountList= accountService.getAccounts();
-		return accountList;
+		accountRegistrationResponse.setAccountList(accountList);
+		accountRegistrationResponse.setSuccess(true);
+		}else {
+			accountRegistrationResponse.setExceptionDetails("Token cannot be validated");
+			accountRegistrationResponse.setSuccess(false);
+		}
+		return accountRegistrationResponse;
 
 	}
 	
 	@GetMapping(value = "/accounts/{accountNum}")
-	public Account getAccountByAccountNum(@PathVariable(value = "accountNum") int accountNum) {
+	public AccountRegistrationResponse getAccountByAccountNum(@PathVariable(value = "accountNum") int accountNum, @RequestHeader String authorization) {
 		logger.error("In get Account by Account Number");
+		
+		AccountRegistrationResponse accountRegistrationResponse = new AccountRegistrationResponse();
+		String isvalidToken = accountService.isTokenValid(authorization);
+		if (isvalidToken.equalsIgnoreCase("success")) {
 		Account account= accountService.getAccountByAccountNum(accountNum);
-		return account;
+		accountRegistrationResponse.setAccount(account);
+		accountRegistrationResponse.setSuccess(true);
+		}else {
+			accountRegistrationResponse.setExceptionDetails("Token cannot be validated");
+			accountRegistrationResponse.setSuccess(false);
+		}
+		return accountRegistrationResponse;
 
 	}
 
 	@PutMapping(value = "/accounts/{accountNum}")
-	public Account updateAccount(@PathVariable(value = "accountNum") int accountNum, @RequestBody AccountRegistration accountInputForm) {
+	public AccountRegistrationResponse updateAccount(@PathVariable(value = "accountNum") int accountNum, @RequestBody AccountRegistration accountInputForm, @RequestHeader String authorization){
 		logger.error("In Update Account");
+		AccountRegistrationResponse accountRegistrationResponse = new AccountRegistrationResponse();
+		String isvalidToken = accountService.isTokenValid(authorization);
+		if (isvalidToken.equalsIgnoreCase("success")) {
 		Account updatedAccount= accountService.updateAccount(accountNum, accountInputForm);
-		return updatedAccount;
+		accountRegistrationResponse.setAccount(updatedAccount);
+		accountRegistrationResponse.setSuccess(true);
+		}else {
+			accountRegistrationResponse.setExceptionDetails("Token cannot be validated");
+			accountRegistrationResponse.setSuccess(false);
+		}
+		return accountRegistrationResponse;
 
 	}
 	
 	@DeleteMapping(value = "/accounts/{accountNum}")
-	public AccountRegistrationResponse deleteAccount(@PathVariable(value = "accountNum") int accountNum) {
+	public AccountRegistrationResponse deleteAccount(@PathVariable(value = "accountNum") int accountNum, @RequestHeader String authorization){
 		logger.error("In Delete Account");
-		AccountRegistrationResponse accountRegistrationResponse = new AccountRegistrationResponse();
 		
+		AccountRegistrationResponse accountRegistrationResponse = new AccountRegistrationResponse();
+		String isvalidToken = accountService.isTokenValid(authorization);
+		if (isvalidToken.equalsIgnoreCase("success")) {
 		boolean isDeleted = accountService.deleteAccount(accountNum);
 		accountRegistrationResponse.setSuccess(isDeleted);
-		accountRegistrationResponse.setAccountNumber(accountNum);
+		accountRegistrationResponse.setAccountNumber(String.valueOf(accountNum));
+		}else {
+			accountRegistrationResponse.setExceptionDetails("Token cannot be validated");
+			accountRegistrationResponse.setSuccess(false);
+		}
 		return accountRegistrationResponse;
 
 	}
